fix api

+++ src/app/api/user/delete/route.ts
@@ -4,14 +4,29 @@ import {NextResponse} from 'next/server';
 import {jwtDecode} from 'jwt-decode';
 
 export async function DELETE (){
+    interface JwtPayload {
+        uid: string;
+        name?: string;
+        email: string;
+    }
     const cookieStore = cookies();
     const secret = process.env.NEXT_PUBLIC_AUTH_SECRET || "";
-    const token = cookieStore.get(process.env.NEXT_PUBLIC_COOKIE_NAME);
+    const token = cookieStore.get(`${process.env.NEXT_PUBLIC_COOKIE_NAME}`);
 
     try {
+        if (!token) {
+            return NextResponse.json(
+                {
+                    message: "Cookie is undefined or has no value",
+                },
+                {
+                    status: 400,
+                }
+            );
+        }
 
         const {value} = token
-        const decoded = jwtDecode(value, secret as any);
+        const decoded = jwtDecode<JwtPayload>(value, secret as any);
 
 
         const result = axios.delete(`${process.env.NEXT_PUBLIC_API_URL}/api/users/${decoded.uid}`, {
@@ -28,7 +43,7 @@ export async function DELETE (){
     }catch (error) {
         return NextResponse.json(
             {
-                message: error.message,
+                message: 'Delete User Failed',
             },
             {
                 status: 400,
